api-automation/requirements.txt
pytest==7.4.3
requests==2.31.0
allure-pytest==2.13.2
pytest-html==4.1.1
pytest-xdist==3.3.1
jsonschema==4.20.0
python-dotenv==1.0.0
faker==20.1.0
pytest-cov==4.1.0
pytest-json-report==1.5.0
pytest-metadata==3.0.0
